Making sure the location function can get the site data.
Using postGIS query functions to either look for the polygon the site is within, or the polygon closest to the object.
If the site is not within another polygon, then we look for the closest polygon. This can be really slow for some sites. Ideally we can spped it up, but `<->` is not an indexed query.