Ading a `diff` and `deactivate` method to the DOIs.
Deactivate should use the `hide` method of the datacite API.
The diff provides a record of things that have changed in the doi metadata. It's a bit ugly. We might be able to generate a better method from scratch at some point.

Sped up the geolocaiton method by using near neighbour on the geometry rather than geography. This does stink a bit if the distance is really big. Should fix.